src/Sign_Language_Classification/utils/common.py
import os
import yaml
import json
import torch
import joblib
from pathlib import Path
from box import ConfigBox
from ensure import ensure_annotations

@ensure_annotations
def read_yaml(path_to_yaml: Path) -> ConfigBox:
    """reads yaml file and returns ConfigBox

    Args:
        path_to_yaml (str): path like input

    Raises:
        ValueError: if yaml file is empty
        e: empty file

    Returns:
        ConfigBox: ConfigBox type
    """
    try:
        with open(path_to_yaml) as yaml_file:
            content = yaml.safe_load(yaml_file)
            return ConfigBox(content)
    except Exception as e:
        raise e

@ensure_annotations
def create_directories(path_to_directories: list, verbose=True):
    """create list of directories

    Args:
        path_to_directories (list): list of path of directories
        verbose (bool, optional): If True log will be shown. Defaults to True.
    """
    for path in path_to_directories:
        os.makedirs(path, exist_ok=True)
        if verbose:
            print(f"created directory at: {path}")

@ensure_annotations
def save_json(path: Path, data: dict):
    """save json data

    Args:
        path (Path): path to json file
        data (dict): data to be saved in json file
    """
    with open(path, "w") as f:
        json.dump(data, f, indent=4)

@ensure_annotations
def load_json(path: Path) -> ConfigBox:
    """load json files data

    Args:
        path (Path): path to json file

    Returns:
        ConfigBox: data as class attributes instead of dict
    """
    with open(path) as f:
        content = json.load(f)

    return ConfigBox(content)

def save_model(path: Path, model):
    """Save PyTorch model

    Args:
        path (Path): path to save model
        model: PyTorch model to save
    """
    torch.save(model.state_dict(), path)

def load_model(path: Path, model):
    """Load PyTorch model

    Args:
        path (Path): path to model file
        model: PyTorch model instance to load weights into

    Returns:
        model: Loaded model
    """
    model.load_state_dict(torch.load(path))
    return model
