requirements.txt
torch
torchvision
transformers
datasets==2.9.0
evaluate
opencv-python
tqdm
pyyaml
matplotlib
scikit-learn
pillow
pytest
dvc
python-box
ensure
joblib
pathlib
imageio
IPython
jupyter
ipykernel
datasets
tf-keras
fastapi
uvicorn
python-multipart